+++ libs/components/action-bars/src/app/public/public_api.ts
@@ -1,1 +1,0 @@
-export * from './modules/summary-action-bar/summary-action-bar.module';